test: simplify fiber registration coverage
